Updated Features Page 1 of 2
Added: shot of header after changes made by the last commit.
Added: boilerplate code for the next page
Added: link of icons, was so pose to use https://github.com/ionic-team/ionicons, but it seams that the team has changed there methods and it not in line with with this video. Will return to Ionicons later next week when working on my new web page.
Added: 4 icons for features page